--- apps/back-office/src/app/components/settings-filter/settings-filter.component.html
@@ -1,45 +1,0 @@
-<form [formGroup]="settingsForm" class="flex gap-4 flex-col">
-   <!-- Filter Variant -->
-  <div uiFormFieldDirective [outline]="false">
-    <label>{{
-      'models.dashboard.filterVariant.title' | translate 
-    }}</label>
-    <ui-select-menu
-          formControlName="variant"
-          [placeholder]="'models.dashboard.filterVariant.variantsOptions.selectOrigin'| translate"
-        >
-          <ui-select-option value="modern">
-            {{ 'models.dashboard.filterVariant.variantsOptions.variant_modern'
-            | translate}}
-          </ui-select-option>
-          <ui-select-option value="default">
-            {{  'models.dashboard.filterVariant.variantsOptions.variant_default'
-            | translate }}
-          </ui-select-option>
-    </ui-select-menu>
-  </div>
-  <!-- Filter structure (survey) -->
-  <ui-button
-    class="!contents"
-    variant="default"
-    (click)="contextService.onEditFilter(); openEditStructure = true"
-  >
-    {{'components.application.dashboard.editFilter' | translate}}
-  </ui-button>
-  <!-- Activate filtering -->
-  <ui-toggle formControlName="showFilter">
-    <ng-container ngProjectAs="label">{{
-      'models.dashboard.activateFiltering' | translate
-    }}</ng-container>
-  </ui-toggle>
-  <!-- Closable status -->
-  <ui-toggle 
-    formControlName="closable"
-    [uiTooltip]="'models.dashboard.tooltip.closable' | translate"
-  >
-    <ng-container ngProjectAs="label">{{
-      'models.dashboard.closable' | translate
-    }}</ng-container>
-  </ui-toggle>
-</form>
-
