+++ apps/back-office/src/app/components/view-settings-modal/view-settings-modal.component.html
@@ -251,10 +251,9 @@
           </ng-container>
           <!-- Tab content -->
           <div class="flex flex-col">
-            <app-settings-filter 
+            <app-dashboard-filter-settings
               [dashboard]="dashboard"
-              (update)="onUpdateDashboardFilter($event)"
-            ></app-settings-filter>
+            ></app-dashboard-filter-settings>
           </div>
         </ui-tab>
       </ui-tabs>
