+++ libs/shared/src/i18n/fr.json
@@ -2093,7 +2093,7 @@
 				"text": "Texte du bouton",
 				"variant": "Variante de bouton"
 			},
-			"closable": "Fermable",
+			"closable": "Peut être fermé",
 			"context": {
 				"datasource": {
 					"alert": "Sélectionnez une source de données, afin de définir une page par enregistrement de la source de données. Vous pouvez également définir un modèle qui sera chargé par défaut si aucune page spécifique n'est créée pour un enregistrement.",
@@ -2133,11 +2133,12 @@
 			"historicalDate": "Récupérer les données historiques",
 			"share": "Partager le tableau de bord",
 			"tooltip": {
-				"closable": "Lorsque la variante est moderne ou que l'application est utilisée comme élément web, elle indiquera si le filtre peut être fermé ou non",
-				"dashboardFilter": "Vous pouvez utiliser les champs du filtre du tableau de bord, ou utiliser des valeurs statiques.",
-				"historicalDate": "Vous pouvez récupérer les données à une date spécifique en utilisant un champ du filtre du tableau de board.\nVeuillez alors utiliser cette syntaxe : {{filter.<nom>}}.",
-				"useDashboardFilterEditor": "Utiliser l'éditeur de tableaux de bord",
-				"useDefaultEditor": "Utiliser l'éditeur par défaut"
+				"filter": {
+					"closable": "Lorsque la variante est moderne ou que l'application est utilisée comme élément web, un bouton s'affichera pour permettre à l'utilisateur de cacher le filtre",
+					"fields": "Vous pouvez utiliser les champs du filtre du tableau de bord, ou utiliser des valeurs statiques.",
+					"position": "Ne s'applique que lorsque la variant par défaut est active"
+				},
+				"historicalDate": "Vous pouvez récupérer les données à une date spécifique en utilisant un champ du filtre du tableau de board.\nVeuillez alors utiliser cette syntaxe : {{filter.<nom>}}."
 			}
 		},
 		"form": {
