--- libs/shared/src/lib/components/dashboard-filter/filter-settings-modal/filter-settings-modal.component.html
@@ -1,27 +1,0 @@
-<ui-dialog>
-  <ng-container ngProjectAs="header">
-    <h3 class="font-semibold">
-      {{
-        'components.application.dashboard.settings.defaultPosition' | translate
-      }}
-    </h3>
-  </ng-container>
-  <ng-container ngProjectAs="content">
-    <div class="flex flex-auto">
-      <ng-container *ngFor="let position of positionList">
-        <ui-button
-          [isIcon]="true"
-          icon="{{ position + '_panel_open' }}"
-          [isOutlined]="true"
-          (click)="setDefaultPosition(position)"
-          [uiTooltip]="positionTooltips[position] | translate"
-        ></ui-button>
-      </ng-container>
-    </div>
-  </ng-container>
-  <ng-container ngProjectAs="actions">
-    <ui-button [uiDialogClose]>
-      {{ 'common.close' | translate }}
-    </ui-button>
-  </ng-container>
-</ui-dialog>
