+++ libs/shared/src/lib/components/filter/filter-row/filter-row.component.html
@@ -25,21 +25,6 @@
     ></ng-container>
   </ng-container>
 
-  <!-- TOGGLE FILTER EDITOR -->
-  <!-- <ui-button
-    *ngIf="field && isFilterEnable"
-    [isIcon]="true"
-    [icon]="isFilterEditorOnView ? 'filter_list_off' : 'filter_list'"
-    [isOutlined]="true"
-    (click)="toggleFilterEditor()"
-    [uiTooltip]="
-      (isFilterEditorOnView
-        ? 'models.dashboard.tooltip.useDefaultEditor'
-        : 'models.dashboard.tooltip.useDashboardFilterEditor'
-      ) | translate
-    "
-  ></ui-button> -->
-
   <!-- DELETE ROW -->
   <ui-button
     [isIcon]="true"
