+++ libs/shared/src/lib/components/widgets/common/contextual-filters-settings/contextual-filters-settings.component.html
@@ -26,7 +26,7 @@
         variant="grey"
         [size]="18"
         uiSuffix
-        [uiTooltip]="'models.dashboard.tooltip.dashboardFilter' | translate"
+        [uiTooltip]="'models.dashboard.tooltip.filter.fields' | translate"
       ></ui-icon>
     </h2>
     <div class="grow overflow-y-auto overflow-x-hidden w-full">
